Fix the messages per second statistic displayed to user
.. when packing a QWK packet (on Win32, was always reported 0 messages per second).

+++ src/sbbs3/pack_qwk.cpp
@@ -575,7 +575,7 @@ bool sbbs_t::pack_qwk(char *packet, ulong *msgcnt, bool prepack)
 				"(%lu messages/second)."
 				,(*msgcnt)+mailmsgs
 				,ftell(qwk)
-				,elapsed
+				,(ulong)elapsed
 				,((*msgcnt)+mailmsgs) / elapsed);
 		lprintf(LOG_INFO, "packed %lu messages (%lu bytes) in %lu seconds (%lu msgs/sec)"
 			,(*msgcnt)+mailmsgs
